LPS-54637 move logic to new IndexBackgroundTaskExecutor from EditServerAction
Conflicts:
	portal-impl/src/com/liferay/portlet/admin/action/EditServerAction.java
