+++ portal-impl/src/com/liferay/portlet/admin/action/EditServerAction.java
@@ -25,13 +25,6 @@ import com.liferay.portal.kernel.cache.MultiVMPoolUtil;
 import com.liferay.portal.kernel.cache.SingleVMPoolUtil;
 import com.liferay.portal.kernel.captcha.Captcha;
 import com.liferay.portal.kernel.captcha.CaptchaUtil;
-import com.liferay.portal.kernel.cluster.ClusterExecutorUtil;
-import com.liferay.portal.kernel.cluster.ClusterNode;
-import com.liferay.portal.kernel.cluster.ClusterRequest;
-import com.liferay.portal.kernel.concurrent.ThreadPoolExecutor;
-import com.liferay.portal.kernel.dao.shard.ShardUtil;
-import com.liferay.portal.kernel.exception.SystemException;
-import com.liferay.portal.kernel.executor.PortalExecutorManagerUtil;
 import com.liferay.portal.kernel.image.GhostscriptUtil;
 import com.liferay.portal.kernel.image.ImageMagickUtil;
 import com.liferay.portal.kernel.io.unsync.UnsyncByteArrayOutputStream;
@@ -42,17 +35,11 @@ import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.log.SanitizerLogWrapper;
 import com.liferay.portal.kernel.mail.Account;
-import com.liferay.portal.kernel.messaging.BaseAsyncDestination;
-import com.liferay.portal.kernel.messaging.Destination;
 import com.liferay.portal.kernel.messaging.DestinationNames;
-import com.liferay.portal.kernel.messaging.MessageBus;
 import com.liferay.portal.kernel.messaging.MessageBusUtil;
-import com.liferay.portal.kernel.messaging.proxy.MessageValuesThreadLocal;
 import com.liferay.portal.kernel.scripting.ScriptingException;
 import com.liferay.portal.kernel.scripting.ScriptingHelperUtil;
 import com.liferay.portal.kernel.scripting.ScriptingUtil;
-import com.liferay.portal.kernel.search.Indexer;
-import com.liferay.portal.kernel.search.IndexerRegistryUtil;
 import com.liferay.portal.kernel.search.SearchEngineUtil;
 import com.liferay.portal.kernel.servlet.DirectServletRegistryUtil;
 import com.liferay.portal.kernel.servlet.SessionErrors;
@@ -60,13 +47,10 @@ import com.liferay.portal.kernel.servlet.SessionMessages;
 import com.liferay.portal.kernel.util.CharPool;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.InstancePool;
-import com.liferay.portal.kernel.util.MethodHandler;
-import com.liferay.portal.kernel.util.MethodKey;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.ProgressStatusConstants;
 import com.liferay.portal.kernel.util.ProgressTracker;
 import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.ThreadUtil;
@@ -74,10 +58,7 @@ import com.liferay.portal.kernel.util.Time;
 import com.liferay.portal.kernel.util.UnsyncPrintWriterPool;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.xuggler.XugglerUtil;
-import com.liferay.portal.search.SearchEngineInitializer;
-import com.liferay.portal.search.lucene.LuceneHelper;
-import com.liferay.portal.search.lucene.LuceneHelperUtil;
-import com.liferay.portal.search.lucene.cluster.LuceneClusterUtil;
+import com.liferay.portal.search.backgroundtask.IndexBackgroundTaskExecutor;
 import com.liferay.portal.security.auth.PrincipalException;
 import com.liferay.portal.security.lang.DoPrivilegedBean;
 import com.liferay.portal.security.membershippolicy.OrganizationMembershipPolicy;
@@ -89,7 +70,9 @@ import com.liferay.portal.security.membershippolicy.SiteMembershipPolicyFactoryU
 import com.liferay.portal.security.membershippolicy.UserGroupMembershipPolicy;
 import com.liferay.portal.security.membershippolicy.UserGroupMembershipPolicyFactoryUtil;
 import com.liferay.portal.security.permission.PermissionChecker;
+import com.liferay.portal.service.BackgroundTaskLocalServiceUtil;
 import com.liferay.portal.service.ServiceComponentLocalServiceUtil;
+import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.struts.ActionConstants;
 import com.liferay.portal.struts.PortletAction;
 import com.liferay.portal.theme.ThemeDisplay;
@@ -98,7 +81,6 @@ import com.liferay.portal.util.MaintenanceUtil;
 import com.liferay.portal.util.Portal;
 import com.liferay.portal.util.PortalInstances;
 import com.liferay.portal.util.PrefsPropsUtil;
-import com.liferay.portal.util.PropsValues;
 import com.liferay.portal.util.ShutdownUtil;
 import com.liferay.portal.util.WebKeys;
 import com.liferay.portlet.ActionResponseImpl;
@@ -107,13 +89,11 @@ import com.liferay.portlet.documentlibrary.util.DLPreviewableProcessor;
 import com.liferay.util.log4j.Log4JUtil;
 
 import java.io.File;
+import java.io.Serializable;
 
 import java.util.Enumeration;
-import java.util.HashSet;
+import java.util.HashMap;
 import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
@@ -375,93 +355,21 @@ public class EditServerAction extends PortletAction {
 	}
 
 	protected void reindex(ActionRequest actionRequest) throws Exception {
+		ThemeDisplay themeDisplay = (ThemeDisplay)actionRequest.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
 		String className = ParamUtil.getString(actionRequest, "className");
+		Map<String, Serializable> taskContextMap = new HashMap<>();
 
 		long[] companyIds = PortalInstances.getCompanyIds();
 
-		if (LuceneHelperUtil.isLoadIndexFromClusterEnabled()) {
-			MessageValuesThreadLocal.setValue(
-				LuceneHelper.SKIP_LOAD_INDEX_FROM_CLUSTER, true);
-		}
+		taskContextMap.put("companyIds", companyIds);
+		taskContextMap.put("className", className);
 
-		Set<String> usedSearchEngineIds = new HashSet<>();
-
-		if (Validator.isNull(className)) {
-			for (long companyId : companyIds) {
-				try {
-					SearchEngineInitializer searchEngineInitializer =
-						new SearchEngineInitializer(companyId);
-
-					searchEngineInitializer.reindex();
-
-					usedSearchEngineIds.addAll(
-						searchEngineInitializer.getUsedSearchEngineIds());
-				}
-				catch (Exception e) {
-					_log.error(e, e);
-				}
-			}
-		}
-		else {
-			Indexer indexer = IndexerRegistryUtil.getIndexer(className);
-
-			if (indexer == null) {
-				return;
-			}
-
-			Set<String> searchEngineIds = new HashSet<>();
-
-			searchEngineIds.add(indexer.getSearchEngineId());
-
-			for (String searchEngineId : searchEngineIds) {
-				for (long companyId : companyIds) {
-					SearchEngineUtil.deleteEntityDocuments(
-						searchEngineId, companyId, className, true);
-				}
-			}
-
-			for (long companyId : companyIds) {
-				ShardUtil.pushCompanyService(companyId);
-
-				try {
-					indexer.reindex(new String[] {String.valueOf(companyId)});
-
-					usedSearchEngineIds.add(indexer.getSearchEngineId());
-				}
-				catch (Exception e) {
-					_log.error(e, e);
-				}
-				finally {
-					ShardUtil.popCompanyService();
-				}
-			}
-		}
-
-		if (!LuceneHelperUtil.isLoadIndexFromClusterEnabled()) {
-			return;
-		}
-
-		Set<BaseAsyncDestination> searchWriterDestinations = new HashSet<>();
-
-		MessageBus messageBus = MessageBusUtil.getMessageBus();
-
-		for (String usedSearchEngineId : usedSearchEngineIds) {
-			String searchWriterDestinationName =
-				SearchEngineUtil.getSearchWriterDestinationName(
-					usedSearchEngineId);
-
-			Destination destination = messageBus.getDestination(
-				searchWriterDestinationName);
-
-			if (destination instanceof BaseAsyncDestination) {
-				BaseAsyncDestination baseAsyncDestination =
-					(BaseAsyncDestination)destination;
-
-				searchWriterDestinations.add(baseAsyncDestination);
-			}
-		}
-
-		submitClusterIndexLoadingSyncJob(searchWriterDestinations, companyIds);
+		BackgroundTaskLocalServiceUtil.addBackgroundTask(
+			themeDisplay.getUserId(), themeDisplay.getScopeGroupId(), "reindex",
+			null, IndexBackgroundTaskExecutor.class, taskContextMap,
+			new ServiceContext());
 	}
 
 	protected void reindexDictionaries(ActionRequest actionRequest)
@@ -537,72 +445,6 @@ public class EditServerAction extends PortletAction {
 				ShutdownUtil.shutdown(minutes, message);
 			}
 		}
-	}
-
-	protected void submitClusterIndexLoadingSyncJob(
-			Set<BaseAsyncDestination> baseAsyncDestinations, long[] companyIds)
-		throws Exception {
-
-		if (_log.isInfoEnabled()) {
-			StringBundler sb = new StringBundler(
-				baseAsyncDestinations.size() + 1);
-
-			sb.append("[");
-
-			for (BaseAsyncDestination baseAsyncDestination :
-					baseAsyncDestinations) {
-
-				sb.append(baseAsyncDestination.getName());
-				sb.append(", ");
-			}
-
-			sb.setStringAt("]", sb.index() - 1);
-
-			_log.info(
-				"Synchronizecluster index loading for destinations " +
-					sb.toString());
-		}
-
-		int totalWorkersMaxSize = 0;
-
-		for (BaseAsyncDestination baseAsyncDestination :
-				baseAsyncDestinations) {
-
-			totalWorkersMaxSize += baseAsyncDestination.getWorkersMaxSize();
-		}
-
-		if (_log.isInfoEnabled()) {
-			_log.info(
-				"There are " + totalWorkersMaxSize +
-					" synchronization threads");
-		}
-
-		CountDownLatch countDownLatch = new CountDownLatch(
-			totalWorkersMaxSize + 1);
-
-		ClusterLoadingSyncJob slaveClusterLoadingSyncJob =
-			new ClusterLoadingSyncJob(companyIds, countDownLatch, false);
-
-		for (BaseAsyncDestination baseAsyncDestination :
-				baseAsyncDestinations) {
-
-			ThreadPoolExecutor threadPoolExecutor =
-				PortalExecutorManagerUtil.getPortalExecutor(
-					baseAsyncDestination.getName());
-
-			for (int i = 0; i < baseAsyncDestination.getWorkersMaxSize(); i++) {
-				threadPoolExecutor.execute(slaveClusterLoadingSyncJob);
-			}
-		}
-
-		ClusterLoadingSyncJob masterClusterLoadingSyncJob =
-			new ClusterLoadingSyncJob(companyIds, countDownLatch, true);
-
-		ThreadPoolExecutor threadPoolExecutor =
-			PortalExecutorManagerUtil.getPortalExecutor(
-				EditServerAction.class.getName());
-
-		threadPoolExecutor.execute(masterClusterLoadingSyncJob);
 	}
 
 	protected void threadDump() throws Exception {
@@ -954,109 +796,5 @@ public class EditServerAction extends PortletAction {
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		EditServerAction.class);
-
-	private static final MethodKey _loadIndexesFromClusterMethodKey =
-		new MethodKey(
-			LuceneClusterUtil.class, "loadIndexesFromCluster", long[].class,
-		ClusterNode.class);
-
-	private static class ClusterLoadingSyncJob implements Runnable {
-
-		public ClusterLoadingSyncJob(
-			long[] companyIds, CountDownLatch countDownLatch, boolean master) {
-
-			_companyIds = companyIds;
-			_countDownLatch = countDownLatch;
-			_master = master;
-		}
-
-		@Override
-		public void run() {
-			_countDownLatch.countDown();
-
-			String logPrefix = StringPool.BLANK;
-
-			if (_log.isInfoEnabled()) {
-				Thread currentThread = Thread.currentThread();
-
-				if (_master) {
-					logPrefix =
-						"Monitor thread name " + currentThread.getName() +
-							" with thread ID " + currentThread.getId();
-				}
-				else {
-					logPrefix =
-						"Thread name " + currentThread.getName() +
-							" with thread ID " + currentThread.getId();
-				}
-			}
-
-			if (!_master && _log.isInfoEnabled()) {
-				_log.info(
-					logPrefix + " synchronized on latch. Waiting for others.");
-			}
-
-			try {
-				if (_master) {
-					_countDownLatch.await();
-				}
-				else {
-					boolean result = _countDownLatch.await(
-						PropsValues.LUCENE_CLUSTER_INDEX_LOADING_SYNC_TIMEOUT,
-						TimeUnit.MILLISECONDS);
-
-					if (!result) {
-						_log.error(
-							logPrefix + " timed out. You may need to " +
-								"re-trigger a reindex process.");
-					}
-				}
-			}
-			catch (InterruptedException ie) {
-				if (_master) {
-					_log.error(
-						logPrefix + " was interrupted. Skip cluster index " +
-							"loading notification.",
-						ie);
-
-					return;
-				}
-				else {
-					_log.error(
-						logPrefix + " was interrupted. You may need to " +
-							"re-trigger a reindex process.",
-						ie);
-				}
-			}
-
-			if (_master) {
-				ClusterRequest clusterRequest =
-					ClusterRequest.createMulticastRequest(
-						new MethodHandler(
-							_loadIndexesFromClusterMethodKey, _companyIds,
-							ClusterExecutorUtil.getLocalClusterNode()),
-						true);
-
-				try {
-					ClusterExecutorUtil.execute(clusterRequest);
-				}
-				catch (SystemException se) {
-					_log.error(
-						"Unable to notify peers to start index loading", se);
-				}
-
-				if (_log.isInfoEnabled()) {
-					_log.info(
-						logPrefix + " unlocked latch. Notified peers to " +
-							"start index loading.");
-				}
-			}
-		}
-
-		private final long[] _companyIds;
-		private final CountDownLatch _countDownLatch;
-		private final boolean _master;
-
-	}
 
 }
